# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  models.CreateEventDTO:
    properties:
      description:
        description: Краткое описание события
        type: string
      event_type:
        allOf:
        - $ref: '#/definitions/models.EventType'
        description: 'тип события: локация, событие, артефакт'
      info:
        description: информация о событии
        type: string
      source:
        description: Источник информации об этом событии
        type: string
      title:
        description: Название события
        type: string
    type: object
  models.EventType:
    enum:
    - Локация
    - Событие
    - Артефакт
    type: string
    x-enum-varnames:
    - EventTypeLocation
    - EventTypeEvent
    - EventTypeArtifact
  models.GetEventsDTO:
    properties:
      events_count:
        type: integer
      historical_events:
        items:
          $ref: '#/definitions/models.HistoricalEvent'
        type: array
      publications_id:
        type: integer
    type: object
  models.HistoricalEvent:
    properties:
      description:
        description: Краткое описание события
        type: string
      event_type:
        allOf:
        - $ref: '#/definitions/models.EventType'
        description: 'тип события: локация, событие, артефакт'
      id:
        description: Уникальный идентификатор события
        type: integer
      info:
        description: информация о событии
        type: string
      photo_url:
        description: URL фотографии, связанной с событием
        type: string
      source:
        description: Источник информации об этом событии
        type: string
      status:
        description: Установлен статус по умолчанию
        type: string
      title:
        description: Название события
        type: string
    type: object
...
  models.UpdateEventDTO:
    properties:
      description:
        description: Краткое описание события
        type: string
      event_type:
        allOf:
        - $ref: '#/definitions/models.EventType'
        description: 'тип события: локация, событие, артефакт'
      info:
        description: информация о событии
        type: string
      source:
        description: Источник информации об этом событии
        type: string
      title:
        description: Название события
        type: string
    type: object
  models.UpdateEventPriority:
    properties:
      priority:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API для управления событиями, публикациями и изображениями.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: RIP API
  version: "1.0"
paths:
  /events:
    get:
      consumes:
      - application/json
      description: Возвращает список событий с возможностью фильтрации по названию.
        Также возвращает черновик публикации пользователя, если он существует.
      parameters:
      - description: Название события
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetEventsDTO'
        "404":
          description: События не найдены
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить список событий
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Добавляет новое событие в систему.
      parameters:
      - description: Информация о событии
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.HistoricalEvent'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Добавить новое событие
      tags:
      - Events
  /events/{eventID}:
    delete:
      consumes:
      - application/json
      description: Удаляет событие по его уникальному идентификатору.
      parameters:
      - description: Идентификатор события
        in: path
        name: eventID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный формат ID события
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Событие не найдено
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Событие уже удалено
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Удалить событие по ID
      tags:
      - Events
    get:
      consumes:
      - application/json
      description: Возвращает событие по уникальному идентификатору.
      parameters:
      - description: Идентификатор события
        in: path
        name: eventID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HistoricalEvent'
        "400":
          description: Неверный формат ID события
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Событие не найдено
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить событие по ID
      tags:
      - Events
    put:
      consumes:
      - application/json
      description: Обновляет информацию о событии по его уникальному идентификатору.
      parameters:
      - description: Идентификатор события
        in: path
        name: eventID
        required: true
        type: integer
      - description: Обновленные данные события
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HistoricalEvent'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Событие не найдено
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Событие уже удалено
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Обновить событие по ID
      tags:
      - Events
  /events/{eventID}/image:
    post:
      consumes:
      - multipart/form-data
      description: Добавляет изображение к событию по его ID.
      parameters:
      - description: Идентификатор события
        in: path
        name: eventID
        required: true
        type: integer
      - description: Изображение события
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Неверный формат ID события или файла
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Событие не найдено
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Добавить изображение к событию
      tags:
      - Events
  /events/{eventID}/priority:
    put:
      consumes:
      - application/json
      description: Обновляет приоритет события в черновике публикации пользователя.
      parameters:
      - description: Идентификатор события
        in: path
        name: eventID
        required: true
        type: integer
      - description: Новый приоритет события
        in: body
        name: priority
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventPriority'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Неверный формат ID события или данных приоритета
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Черновик публикации не найден
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Публикация уже удалена
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Обновить приоритет события
      tags:
      - Events-Publications
  /events/{eventID}/publications:
    delete:
      consumes:
      - application/json
      description: Удаляет событие из черновика публикации пользователя.
      parameters:
      - description: Идентификатор события
        in: path
        name: eventID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный формат ID события
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Событие или публикация не найдены
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Удалить событие из публикации
      tags:
      - Events-Publications
    post:
      consumes:
      - application/json
      description: Добавляет событие в черновик публикации пользователя.
      parameters:
      - description: Идентификатор события
        in: path
        name: eventID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный формат ID события
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Событие или черновик публикации не найдено
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Добавить событие в публикацию
      tags:
      - Events
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Use the JWT Bearer Token for authorization. Example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"